feat: Order FindAll, Update 기능 구현

+++ src/main/java/flight_booking/demo/domain/order/dto/request/OrderUpdateRequestDto.java
@@ -4,8 +4,7 @@ import flight_booking.demo.domain.order.entity.OrderState;
 import jakarta.persistence.*;
 
 public record OrderUpdateRequestDto (
-        int ticketId,
-        //@Convert(converter = OrderState.Convertor.class)
+        long ticketId,
         OrderState orderState
 ){
 }
